+++ source/PluginProcessor.h
@@ -58,7 +58,7 @@ public:
 
 private:
     size_t fft_size;
-    int hop_size;
+    size_t hop_size;
     float sample_rate;
 
     struct fftw_data
